Finish sorting materials, Added material registrations
