+++ groovy/preInit/MaterialChanges.groovy
@@ -1,3 +1,4 @@
+/*
 import net.minecraftforge.fml.common.eventhandler.EventPriority
 import gregtech.api.GregTechAPI
 
@@ -8,7 +9,7 @@ import classes.ChangeFlags
 
 log.infoMC("Registering new material event listener")
 
-eventManager.listen(/*EventPriority.LOWEST*/) {
+eventManager.listen(EventPriority.LOWEST) {
     GregTechAPI.MaterialEvent event ->
 
         log.infoMC("Starting new material registration")
@@ -22,3 +23,4 @@ eventManager.listen(/*EventPriority.LOWEST*/) {
         ChangeFlags.init()
 
 }
+*/
